requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pydantic-core>=2.18.0

# Astrological Calculations (CORRECTED)
pyswisseph==2.10.3.2

# Chinese Calendar (Saju/Bazi)
lunar-python>=1.0.0

# Geocoding
geopy>=2.4.0

# Graph Database
neo4j>=5.18.0

# LLM Integration
openai>=1.30.0

# CLI
click>=8.1.0

# Utilities
python-dotenv>=1.0.0
h11~=0.16.0
pip~=26.0.1
typing_extensions~=4.15.0
distro~=1.9.0
protobuf~=6.33.5
cryptography~=46.0.5
pytz~=2025.2
requests~=2.32.5
PyYAML~=6.0.3
sniffio~=1.3.1
geographiclib~=2.1
httpcore~=1.0.9
idna~=3.11
httpx~=0.28.1
websockets~=16.0
anyio~=4.12.1
pydantic_core~=2.41.5
uvloop~=0.22.1
certifi~=2026.2.25
starlette~=0.36.3
httptools~=0.7.1
watchfiles~=1.1.1
annotated-types~=0.7.0
lunar_python~=1.0.0
typing_extensions~=4.15.0
pydantic_core~=2.41.5
lunar_python~=1.0.0